+++ client/src/hooks/index.ts
@@ -1,3 +1,5 @@
+export * from './Messages';
+
 export * from './AuthContext';
 export * from './ThemeContext';
 export * from './ScreenshotContext';
